requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-crispy-forms==1.14.0
idna==3.3
Pillow==9.1.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==3.2.0
twilio==7.9.1
urllib3==1.26.9
